requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
phonenumbers
opencage
folium
python-dotenv
SQLAlchemy
werkzeug